slmk: Don't factor in the size of the processes when choosing victims
With the introduction of the reaper thread as well as fixes to the task
killing code, memory can be freed fast enough without causing lockups
even when we don't prioritize killing processes with a lower adj over a
smaller process with a higher adj and end up killing more processes.

This patch allows more unneeded apps to be killed and more memory to be
freed, which should hopefully decrease lmk wakeups in the long term, and
allow memory to be used for more important apps.

Signed-off-by: Kazuki H <kazukih0205@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>